log4tango: Reimplement threading utilities with standard C++11 features
We don't need to have platform specific solutions with pthread and
Windows threads. Instead we can just use the default
std::mutex, std::recursive_mutex and std::lock_guard classes.
