+++ log4tango/config/config.cmake
@@ -27,13 +27,6 @@ endmacro()
 LOG4TANGO_CHECK_FUNCTION_EXISTS(ftime         HAVE_FTIME)
 LOG4TANGO_CHECK_FUNCTION_EXISTS(gettimeofday  HAVE_GETTIMEOFDAY)
 
-
-#check threads
-check_library_exists(pthread pthread_attr_get_np "" HAVE_PTHREAD_ATTR_GET_NP)
-check_library_exists(pthread pthread_getattr_np "" HAVE_PTHREAD_GETATTR_NP)
-check_library_exists(pthread pthread_create "" HAVE_PTHREAD_CREATE)
-check_include_file("pthread.h;pthread_np.h" HAVE_PTHREAD_NP_H)
-
 if(CMAKE_THREAD_LIBS_INIT)
     set(LOG4TANGO_HAVE_THREADING "/**/")
 endif()
