+++ log4tango/src/CMakeLists.txt
@@ -19,8 +19,7 @@ set(SOURCES
                 StringUtil.hh
                 StringUtil.cpp
                 DummyThreads.cpp
-                MSThreads.cpp
-                PThreads.cpp
+                StandardThreads.cpp
                 PortabilityImpl.hh
                 PortabilityImpl.cpp)
